refactored the reducer
