+++ src/components/Question.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { fetchLevelQuestions } from '../store';
+import { fetchLevelQuestions } from '../store/reducers/currentLevelQuestions';
 import MultipleChoice from './MultipleChoice';
 import ProgressBar from './ProgressBar/ProgressBar';
 import InfoCard from './InfoCard';
@@ -59,7 +59,7 @@ class Question extends Component {
         question.type === 'infoCard')
     ) {
       return (
-        <div className='card question'>
+        <div className="card question">
           <ProgressBar progress={this.state.percentage} />
           {question.type === 'infoCard' ? (
             <InfoCard
